Array Deletion
Two friends are playing a puzzle. They have set of sticks to arrange the puzzle.
They need to delete one puzzle to make the correct puzzle.
So for making correct puzzle you have to find the position where we have to move off the stick.
Help them to correct the puzzle.

Sample Input and Output :

Enter the number of elements in the array
5

Enter the elements in the array
1
2
3
4
5

Enter the location where you wish to delete an element
4

Array after deletion is
1
2
3
5