Array Mean
Write a program to find the mean of the elements in the array.

Input and Output Format:
Input consists of n+1 integers where n corresponds to the number of elements in the array.
The first integer corresponds to n and the next n integers correspond to the elements in the array.

Output consists of a double value which corresponds to the mean of the array. It is printed upto 2 digits of precision.
Assume that the maximum number of elements in the array is 20.

Sample Input and Output :

Enter the number of elements in the array
5

Enter the elements in the array
2
4
1
3
5

The mean of the array is 3.00